[MOB-45] Add custom decorators shared file
