+++ src/infrastructure/http/dtos/update-event.dto.ts
@@ -15,9 +15,14 @@ import {
   MinLength,
   ValidateIf,
   ValidateNested,
-  ValidationOptions,
-  registerDecorator,
 } from 'class-validator';
+
+import {
+  IsValidDate,
+  IsNotPastDate,
+  IsNotEmptyArray,
+  IsNotEmptyObject,
+} from './custom-decorators';
 
 export enum UpdateEventType {
   PARTY = 'Party',
@@ -25,94 +30,6 @@ export enum UpdateEventType {
   FOOD = 'Food',
   CHARITY = 'Charity',
   GAMES = 'Games',
-}
-
-function isValidISODate(eventDate: string): boolean {
-  const [datePart, timePartWithZ] = eventDate.split('T');
-  const timePart = timePartWithZ.replace('Z', '');
-  const [year, month, day] = datePart.split('-').map(Number);
-  const [hour, minute, second, millisecond] = timePart
-    .split(/[:.]/)
-    .map(Number);
-  const date = new Date(
-    Date.UTC(year, month - 1, day, hour, minute, second, millisecond),
-  );
-
-  return (
-    date.getUTCFullYear() === year &&
-    date.getUTCMonth() === month - 1 &&
-    date.getUTCDate() === day &&
-    date.getUTCHours() === hour &&
-    date.getUTCMinutes() === minute &&
-    date.getUTCSeconds() === second &&
-    date.getUTCMilliseconds() === millisecond
-  );
-}
-
-function IsValidDate(validationOptions?: ValidationOptions) {
-  return function (target: object, propertyName: string) {
-    registerDecorator({
-      target: target.constructor,
-      propertyName: propertyName,
-      options: validationOptions,
-      constraints: [],
-      validator: {
-        validate(value: any) {
-          return typeof value === 'string' && isValidISODate(value);
-        },
-      },
-    });
-  };
-}
-
-function IsNotPastDate(validationOptions?: ValidationOptions) {
-  return function (target: object, propertyName: string) {
-    registerDecorator({
-      target: target.constructor,
-      propertyName: propertyName,
-      options: validationOptions,
-      validator: {
-        validate(value: any) {
-          const date = new Date(value);
-          return date > new Date();
-        },
-      },
-    });
-  };
-}
-
-export function IsNotEmptyObject(validationOptions?: ValidationOptions) {
-  return function (object: object, propertyName: string) {
-    registerDecorator({
-      target: object.constructor,
-      propertyName: propertyName,
-      options: validationOptions,
-      validator: {
-        validate(value: any) {
-          return (
-            typeof value === 'object' &&
-            value !== null &&
-            Object.keys(value).length > 0
-          );
-        },
-      },
-    });
-  };
-}
-
-export function IsNotEmptyArray(validationOptions?: ValidationOptions) {
-  return function (object: object, propertyName: string) {
-    registerDecorator({
-      target: object.constructor,
-      propertyName: propertyName,
-      options: validationOptions,
-      validator: {
-        validate(value: any) {
-          return Array.isArray(value) && value !== null && value.length > 0;
-        },
-      },
-    });
-  };
 }
 
 class UpdateEventLocationDto {
@@ -201,8 +118,8 @@ export class UpdateEventDto {
 
   @ApiPropertyOptional({ type: UpdateEventLocationDto })
   @IsNotEmpty()
-  @IsObject()
   @IsNotEmptyObject({ message: 'location should not be empty' })
+  @IsObject()
   @ValidateNested()
   @Type(() => UpdateEventLocationDto)
   @IsOptional()
